# src/ZepelimAuth.Api/ViewModels/OnlyIdViewModel.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace FanPush.ViewModel
{
    public class OnlyIdViewModel
    {
        public int Id { get; set; }
    }
}